Changed the PID to temperature calculation. Previously on a boiler that (#4)
has a min flow temp of 50C and a max of 70C set the pid output would be
clipped, a pid output of 100% would be 70C as you expect, but a pid
output of 70% is the same as 100% - 70C. The same goes for minimum flow
temp, 1% to 50% pid output produces a flow temp of 50C.

This is problematic for PID tuning, producing results which dont work.
It also causes bigger jumps in temperature when PID is modulating in the
50-70% range, a large drop in requested temperature will result in the
boiler shutting off instead of modulating down. Since I implemented the
change on my system I'm seeing more stable temperature regulation,
much less target temperature overshoot, and much less boiler cycling on
and off, which is how an opentherm system should work. Setting output
sample averaging also helps to smooth the boiler setpoint, I am
currently using 2.

My system is an Intergas 18kw system boiler with 14 radiators in a poorly
insulated Victorian home built in 1898 for reference. The outside
temperature while testing has been in the range -4 to 2C, inside room
setpoint 20.8C